msm: mdss: Allocate correct space for hist bin count in 64 bit
Histogram bin count variable was being allocated insufficient
space in compat ioctl for histogram read.
Correcting the space being allocated to the member variable
bin_cnt in histogram read.

+++ drivers/video/fbdev/msm/mdss_compat_utils.c
@@ -2140,7 +2140,7 @@ static int __from_user_hist_data(
 			sizeof(uint32_t)) ||
 	    copy_in_user(&hist_data->bin_cnt,
 			&hist_data32->bin_cnt,
-			sizeof(uint8_t)))
+			sizeof(uint32_t)))
 		return -EFAULT;
 
 	if (get_user(data, &hist_data32->c0) ||
@@ -2167,7 +2167,7 @@ static int __to_user_hist_data(
 			sizeof(uint32_t)) ||
 	    copy_in_user(&hist_data32->bin_cnt,
 			&hist_data->bin_cnt,
-			sizeof(uint8_t)))
+			sizeof(uint32_t)))
 		return -EFAULT;
 
 	if (get_user(data, (unsigned long *) &hist_data->c0) ||
